Update README.md
There is a lot of material.  
Some of them are marked with ***** which me and some of my colleagues-users consider to rank good to excellent.
I am planning on going through them to find references, quizzes, tests, projects, etc.
There are probably some duplicates.  I have not cleaned it up completely.
They are not in any specific order of importance either.
Please, add some courses and of yours, so everyone can use them for our mutual benefit!